Edit deleteCol event listener to remove object
The event listener for deleteCol wasn't removing the object asociated
with the row from myLibrary array, so adding another item after deleting
one restored ALL objects to the table. Edit deleteCol event listener so
that it removes the table display and also the object so that future
additions don't add previously deleted objects.